src/requirements.txt
timm
matplotlib
tqdm
transformers
pytorch-ignite